Updated ShazHack with modularity and 0.8.X compatibility
Updated ShazHack with modularity and 0.8.X compatibility
