+++ module/helpers/config.js
@@ -1,0 +1,19 @@
+export const SHAZHACK = {};
+
+/**
+ * The set of Ability Scores used within the sytem.
+ * @type {Object}
+ */
+ SHAZHACK.attributes = {
+  "phys": "SHAZHACK.AbilityPhys",
+  "agil": "SHAZHACK.AbilityAgil",
+  "intu": "SHAZHACK.AbilityIntu",
+  "pres": "SHAZHACK.AbilityPres"
+};
+
+SHAZHACK.attributesAbbreviations = {
+  "phys": "SHAZHACK.AbilityPhysAbbr",
+  "agil": "SHAZHACK.AbilityAgilAbbr",
+  "intu": "SHAZHACK.AbilityIntuAbbr",
+  "pres": "SHAZHACK.AbilityPresAbbr"
+};
