+++ module/item.js
@@ -1,0 +1,30 @@
+/**
+ * Extend the basic Item with some very simple modifications.
+ * @extends {Item}
+ */
+export class ShazHackItem extends Item {
+  /**
+   * Augment the basic Item data model with additional dynamic data.
+   */
+  prepareData() {
+    super.prepareData();
+
+    // Get the Item's data
+    const itemData = this.data;
+    const actorData = this.actor ? this.actor.data : {};
+
+  }
+
+  /**
+   * Handle clickable rolls.
+   * @param {Event} event   The originating click event
+   * @private
+   */
+  async roll() {
+    // Basic template rendering data
+    const token = this.actor.token;
+    const item = this.data;
+    const actorData = this.actor ? this.actor.data.data : {};
+    const itemData = item.data;
+  }
+}
