--- scss/components/_forms.scss
@@ -1,74 +1,0 @@
-.item-form {
-  font-family: $font-primary;
-}
-
-.sheet-header {
-  flex: 0 0 260px;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  margin-bottom: 10px;
-
-  .profile-img {
-    object-fit: contain;
-    flex: 0 0 200px;
-    height: 200px;
-    margin-right: 10px;
-    border: 0px solid #000;
-  }
-
-  .profile-img-item {
-    object-fit: contain;
-    vertical-align: top;
-    flex: 0 0 100px;
-    height: 100px;
-    margin-right: 10px;
-    border: 0px solid #000;
-  }
-
-  .roll-img {
-    object-fit: contain;
-    vertical-align: top;
-    flex: 0 0 20px;
-    height: 20px;
-    border: 0.5px solid #000;
-  }
-
-  .header-fields {
-    flex: 1;
-  }
-
-  h1.charname {
-    height: 50px;
-    padding: 0px;
-    margin: 5px 0;
-    border-bottom: 0;
-    input {
-      width: 100%;
-      height: 100%;
-      margin: 0;
-    }
-  }
-}
-
-.sheet-tabs {
-  flex: 0;
-}
-
-.sheet-body,
-.sheet-body .tab,
-.sheet-body .tab .editor {
-  height: 100%;
-}
-
-.tox {
-  .tox-editor-container {
-    background: $c-white;
-  }
-
-  .tox-edit-area {
-    padding: 0 8px;
-  }
-}
