--- scss/components/_items.scss
@@ -1,46 +1,0 @@
-.items-list {
-  list-style: none;
-  margin: 7px 0;
-  padding: 0;
-  overflow-y: auto;
-
-  .item-header {
-    font-weight: bold;
-  }
-
-  .item {
-    height: 30px;
-    line-height: 24px;
-    padding: 3px 0;
-    border-bottom: 1px solid #BBB;
-
-    .item-image {
-      flex: 0 0 24px;
-      margin-right: 5px;
-    }
-
-    img {
-      display: block;
-    }
-  }
-
-  .item-name {
-    margin: 0;
-    min-width: 40%;
-  }
-
-  .item-name-weapon {
-    margin: 0;
-    min-width: 20%;
-  }
-
-  .item-name-level {
-    margin: 0;
-    max-width: 14%;
-  }
-
-  .item-controls {
-    flex: 0 0 86px;
-    text-align: right;
-  }
-}
