--- scss/components/_tabs.scss
@@ -1,34 +1,0 @@
-// .tabs {
-//   .item {
-//     background: rgba(0, 0, 0, 0.125);
-//     border: 1px solid;
-//     border-radius: 4px;
-//     margin: 0 10px 0;
-//     padding: 8px 4px;
-
-//     &:hover,
-//     &:focus {
-//       background: rgba(0, 0, 0, 0.125);
-//     }
-
-//     &.active {
-//       background: transparent;
-//     }
-//   }
-// }
-
-.tabs {
-  height: 40px;
-  border-top: 1px solid #AAA;
-  border-bottom: 1px solid #AAA;
-
-  .item {
-    line-height: 40px;
-    font-weight: bold;
-  }
-
-  .item.active {
-    text-decoration: underline;
-    text-shadow: none;
-  }
-}
