--- scss/global/_flex.scss
@@ -1,9 +1,0 @@
-.flex-center {
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-}
-
-.flex-between {
-  justify-content: space-between;
-}
