--- scss/global/_window.scss
@@ -1,13 +1,0 @@
-.window-app {
-  font-family: $font-primary;
-}
-
-.rollable {
-  &:hover,
-  &:focus {
-    color: #000;
-    text-shadow: 0 0 10px red;
-    box-shadow: 0 0 5px red;
-    cursor: pointer;
-  }
-}
