--- scss/shazhack.scss
@@ -1,18 +1,0 @@
-// Import utilities.
-@import 'utils/typography';
-@import 'utils/colors';
-@import 'utils/mixins';
-@import 'utils/variables';
-
-/* Global styles */
-@import 'global/window';
-@import 'global/grid';
-@import 'global/flex';
-
-/* Styles limited to shazhack sheets */
-.shazhack {
-  @import 'components/forms';
-  @import 'components/resource';
-  @import 'components/tabs';
-  @import 'components/items';
-}
