--- scss/utils/_colors.scss
@@ -1,2 +1,0 @@
-$c-white: #fff;
-$c-black: #000;
