--- scss/utils/_mixins.scss
@@ -1,16 +1,0 @@
-@mixin element-invisible {
-  position: absolute;
-
-  width: 1px;
-  height: 1px;
-  margin: -1px;
-  border: 0;
-  padding: 0;
-
-  clip: rect(0 0 0 0);
-  overflow: hidden;
-}
-
-@mixin hide {
-  display: none;
-}
