--- scss/utils/_typography.scss
@@ -1,4 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Martel:wght@400;800&family=Roboto:wght@300;400;500&display=swap');
-
-$font-primary: 'Roboto', sans-serif;
-$font-secondary: 'Roboto', sans-serif;
