--- scss/utils/_variables.scss
@@ -1,3 +1,0 @@
-$padding-sm: 5px;
-$padding-md: 10px;
-$padding-lg: 20px;
