+++ templates/character-sheet.hbs
@@ -1,0 +1,23 @@
+<form class="{{cssClass}} flexcol" autocomplete="off">
+  <header class="sheet-header">
+    {{> "systems/shazhack/templates/partials/character-header.hbs"}}
+  </header>
+  {{!-- Sheet Tab Navigation --}}
+  <nav class="sheet-tabs tabs" data-group="primary">
+    <a class="item" data-tab="description">Character</a>
+    <a class="item" data-tab="items">Inventory</a>
+    <a class="item" data-tab="feats">Feats</a>
+    <a class="item" data-tab="spheres">Spheres</a>
+  </nav>
+  {{!-- Sheet Body --}}
+  <section class="sheet-body">
+    {{!-- Biography Tab --}}
+    {{> "systems/shazhack/templates/partials/description-sheet.hbs"}}
+    {{!-- Owned Items Tab --}}
+    {{> "systems/shazhack/templates/partials/items-sheet.hbs"}}
+    {{!-- Owned Features Tab --}}
+    {{> "systems/shazhack/templates/partials/feats-sheet.hbs"}}
+    {{!-- Owned Spheres Tab --}}
+    {{> "systems/shazhack/templates/partials/spheres-sheet.hbs"}}
+  </section>
+</form>
