+++ templates/partials/character-header.hbs
@@ -1,0 +1,30 @@
+<img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" height="210" width="210" />
+<div class="header-fields">
+  <h1 class="charname"><input name="name" type="text" value="{{actor.name}}" placeholder="Name" /></h1>
+  {{!-- The grid classes are defined in scss/global/_grid.scss. To use,
+  use both the "grid" and "grid-Ncol" class where "N" can be any number
+  from 1 to 12 and will create that number of columns. --}}
+  <div class="resources grid grid-2col">
+    {{!-- "flex-group-center" is also defined in the _grid.scss file
+    and it will add a small amount of padding, a border, and will
+    center all of its child elements content and text. --}}
+    <div class="resource flex-group-center">
+      {{> "systems/shazhack/templates/partials/header-partials/hitpoints-partial.hbs"}}
+    </div>
+    <div class="resource flex-group-center">
+      {{> "systems/shazhack/templates/partials/header-partials/hitdice-partial.hbs"}}
+    </div>
+    {{!-- The grid classes are defined in scss/global/_grid.scss. To use,
+    use both the "grid" and "grid-Ncol" class where "N" can be any number
+    from 1 to 12 and will create that number of columns. --}}
+    <div class="attributes grid grid-2col">
+      <div class="resource flex-group-center">
+        {{> "systems/shazhack/templates/partials/header-partials/attributes-partial.hbs"}}
+      </div>
+      <div class="resource flex-group-center">
+        {{> "systems/shazhack/templates/partials/header-partials/attackbonus-partial.hbs"}}
+        {{> "systems/shazhack/templates/partials/header-partials/encumbrance-partial.hbs"}}
+      </div>
+    </div>
+  </div>
+</div>
