+++ templates/partials/character-partials/backgrounds-partial.hbs
@@ -1,0 +1,27 @@
+<label class="resource-label">Backgrounds</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-image"></div>
+    <div class="item-name">Name</div>
+    <div class="item-name-level" text-align: right>Level</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Background"><i class="fas fa-plus"></i>
+        Add
+      </a>
+    </div>
+  </li>
+  {{#each backgrounds as |item id|}}
+  <li class="item flexrow" data-item-id="{{item._id}}">
+    <div class="item-image rollable-background" padding-left: 2px; data-roll="d20+{{item.data.level}}d6kh1"
+      data-label="{{item.name}}">
+      <img src="icons/svg/d20-highlight.svg" title="{{item.name}}" width="24" height="24" />
+    </div>
+    <h4 class="item-name">{{item.name}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.level}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
