+++ templates/partials/character-partials/esoterica-partial.hbs
@@ -1,0 +1,20 @@
+<label class="resource-label">Esoterica</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-name">Name</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Esoterica"><i class="fas fa-plus"></i>
+        Add
+      </a>
+    </div>
+  </li>
+  {{#each esoterica as |item id|}}
+  <li class="item flexrow" data-item-id="{{item._id}}">
+    <h4 class="item-name">{{item.name}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
