+++ templates/partials/feats-sheet.hbs
@@ -1,0 +1,24 @@
+    <div class="tab feats" data-group="primary" data-tab="feats">
+      <ol class="items-list">
+        <li class="item flexrow item-header">
+          <div class="item-name">Name</div>
+          <div class="item-name-level" text-align: right>Choice</div>
+          <div class="item-name-level" text-align: right>Level</div>
+          <div class="item-controls">
+            <a class="item-control item-create" title="Create item" data-type="Feat"><i class="fas fa-plus"></i> Add
+            </a>
+          </div>
+        </li>
+        {{#each feats as |item id|}}
+        <li class="item flexrow" data-item-id="{{item._id}}">
+          <h4 class="item-name">{{item.name}}</h4>
+          <h4 class="item-name-level">{{item.data.choice}}</h4>
+          <h4 class="item-name-level">{{item.data.level}}</h4>
+          <div class="item-controls">
+            <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+            <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+          </div>
+        </li>
+        {{/each}}
+      </ol>
+    </div>
