+++ templates/partials/header-partials/attackbonus-partial.hbs
@@ -1,0 +1,8 @@
+<div class="ability flexrow flex-group-center">
+  <label for="data.combat.attack.value" class="resource-label-noroll">Attack Bonus:</label>
+  <div class="resource-content flexrow flex-center flex-between">
+    <span class="resource-input"></span>
+    <input class="resource-input" readonly type="text" name="data.combat.attack.value"
+      value="{{actor.data.data.combat.attack.value}}" data-dtype="Number" />
+  </div>
+</div>
