+++ templates/partials/header-partials/encumbrance-partial.hbs
@@ -1,0 +1,10 @@
+<div class="ability flexrow flex-group-center">
+  <label for="data.encumbrance.value" class="resource-label-noroll">Encumbrance:</label>
+  <div class="resource-content flexrow flex-center flex-between">
+    <input class="resource-input" type="text" name="data.encumbrance.value"
+      value="{{actor.data.data.encumbrance.value}}" data-dtype="Number" />
+    <span>/</span>
+    <input class="resource-input" type="text" readonly name="data.encumbrance.max"
+      value="{{actor.data.data.encumbrance.max}}" data-dtype="Number" />
+  </div>
+</div>
