+++ templates/partials/header-partials/hitpoints-partial.hbs
@@ -1,0 +1,8 @@
+<div class="resource-content flexrow flex-center flex-between">
+  <label for="actor.data.hitPoints.value" class="resource-label-noroll">Total Hit Points:</label>
+  <input type="text" class="resource-input" name="data.hitPoints.value" value="{{data.hitPoints.value}}"
+    data-dtype="Number" />
+  <span>/</span>
+  <input type="text" class="resource-input" name="data.hitPoints.max" value="{{data.hitPoints.max}}"
+    data-dtype="Number" />
+</div>
