+++ templates/partials/item-partials/armour-partial.hbs
@@ -1,0 +1,30 @@
+<label class="resource-label">Armour</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-image"></div>
+    <div class="item-image"></div>
+    <div class="item-name">Name</div>
+    <div class="item-name-level" text-align: right>Type</div>
+    <div class="item-name-level" text-align: right>Armour Die</div>
+    <div class="item-name-level" text-align: right>Enc.</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Armour"><i class="fas fa-plus"></i> Add
+      </a>
+    </div>
+  </li>
+  {{#each armour as |item id|}}
+  <li class="item flexrow" data-item-id="{{item._id}}">
+    <img class="item-image rollable-armour" src="icons/svg/d20-highlight.svg" data-label="{{item.name}}"
+      title="{{item.name}}" width="24" height="24" data-roll="{{item.data.armourDie}}" />
+    <div class="item-image"><img src="{{item.img}}" title="{{item.name}}" width="24" height="24" /></div>
+    <h4 class="item-name">{{item.name}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.type}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.armourDie}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.encumbrance}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
