+++ templates/partials/item-partials/equipment-partial.hbs
@@ -1,0 +1,31 @@
+<label class="resource-label">Equipment</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-image"></div>
+    <div class="item-image"></div>
+    <div class="item-name">Name</div>
+    <div class="item-name-level" text-align: right>Quantity</div>
+    <div class="item-name-level" text-align: right>Usage Die</div>
+    <div class="item-name-level" text-align: right>Enc.</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Equipment"><i class="fas fa-plus"></i>
+        Add
+      </a>
+    </div>
+  </li>
+  {{#each equipment as |item id|}}
+  <li class="item flexrow" data-item-id="{{item._id}}">
+    <img class="item-image rollable" src="icons/svg/d20-highlight.svg" title="{{item.name}}" width="24" height="24"
+      data-label="{{item.name}}" data-roll="{{item.data.usage}}" />
+    <div class="item-image"><img src="{{item.img}}" title="{{item.name}}" width="24" height="24" /></div>
+    <h4 class="item-name">{{item.name}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.quantity}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.usage}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.encumbrance}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
