+++ templates/partials/sphere-partials/spell-partial.hbs
@@ -1,0 +1,30 @@
+<label class="resource-label">Spells</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-image"></div>
+    <div class="item-image"></div>
+    <div class="item-name-weapon">Name</div>
+    <div class="item-name-weapon" text-align: right>Sphere</div>
+    <div class="item-name-level" text-align: right>Cost</div>
+    <div class="item-name-level" text-align: right>Dice</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Spell"><i class="fas fa-plus"></i> Add</a>
+    </div>
+  </li>
+  {{#each spells as |item id|}}
+  <li class="item flexrow" data-item-id="{{item._id}}">
+    <img class="item-image rollable-spell" src="icons/svg/d20-highlight.svg" title="{{item.name}}" width="24"
+      height="24" data-label="{{item.name}}" data-sphere="{{item.data.sphere}}" data-roll="{{item.data.dice}}"
+      data-cost="{{item.data.cost}}" data-type="{{item.data.type}}" />
+    <div class="item-image"><img src="{{item.img}}" title="{{item.name}}" width="24" height="24" /></div>
+    <h4 class="item-name-weapon">{{item.name}}</h4>
+    <h4 class="item-name-weapon" text-align: right>{{item.data.sphere}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.cost}}</h4>
+    <h4 class="item-name-level" text-align: right>{{item.data.dice}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
