+++ templates/partials/sphere-partials/sphere-partial.hbs
@@ -1,0 +1,27 @@
+<label class="resource-label">Spheres</label>
+<ol class="items-list">
+  <li class="item flexrow item-header">
+    <div class="item-image"></div>
+    <div class="item-image"></div>
+    <div class="item-name">Name</div>
+    <div class="item-name-level" text-align: right>Level</div>
+    <div class="item-name-level" text-align: right>Cost Red.</div>
+    <div class="item-controls">
+      <a class="item-control item-create" title="Create item" data-type="Sphere"><i class="fas fa-plus"></i> Add</a>
+    </div>
+  </li>
+  {{#each spheres}}
+  <li class="item flexrow" data-item-id="{{this._id}}">
+    <img class="item-image rollable-sphere" src="icons/svg/d20-highlight.svg" title="{{item.name}}" width="24"
+      height="24" data-label="{{this.name}}" data-reduction="{{this.data.costReduction}}" />
+    <div class="item-image"><img src="{{this.img}}" title="{{this.name}}" width="24" height="24" /></div>
+    <h4 class="item-name" title="{{actor.name}}">{{this.name}}</h4>
+    <h4 class="item-name-level">{{this.data.level}}</h4>
+    <h4 class="item-name-level">{{this.data.costReduction}}</h4>
+    <div class="item-controls">
+      <a class="item-control item-edit" title="Edit Item"><i class="fas fa-edit"></i></a>
+      <a class="item-control item-delete" title="Delete Item"><i class="fas fa-trash"></i></a>
+    </div>
+  </li>
+  {{/each}}
+</ol>
