+++ templates/partials/spheres-sheet.hbs
@@ -1,0 +1,4 @@
+<div class="tab spheres" data-group="primary" data-tab="spheres">
+  {{> "systems/shazhack/templates/partials/sphere-partials/sphere-partial.hbs"}}
+  {{> "systems/shazhack/templates/partials/sphere-partials/spell-partial.hbs"}}
+</div>
